added XYZPointReader; removed warnings

+++ PotreeConverter/include/PointAttributes.hpp
@@ -26,36 +26,9 @@ public:
 		this->byteSize = byteSize;
 	}
 
-	//bool operator==(const PointAttribute &other){
-	//	return ordinal == other.ordinal;
-	//}
-
 };
 
 bool operator==(const PointAttribute& lhs, const PointAttribute& rhs);
-
-//class PointAttributeDef{
-//public:
-//	string name;
-//	int numElements;
-//	int byteSize;
-//
-//	PointAttributeDef(string name, int numElements, int byteSize){
-//		this->name = name;
-//		this->numElements = numElements;
-//		this->byteSize = byteSize;
-//	}
-//};
-//
-//namespace PointAttribute{
-//	PointAttributeDef POSITION_CARTESIAN("POSITION_CARTESIAN", 3, 3*sizeof(float));
-//	PointAttributeDef COLOR_PACKED("COLOR_PACKED", 4, 4*sizeof(unsigned char));
-//}
-
-//struct PointAttribute{
-//	const PointAttributeDef POSITION_CARTESIAN("POSITION_CARTESIAN", 3, 3*sizeof(float));
-//	const PointAttributeDef COLOR_PACKED("COLOR_PACKED", 4, 4*sizeof(unsigned char));
-//};
 
 
 class PointAttributes{
@@ -73,7 +46,7 @@ public:
 	}
 
 	int size(){
-		return attributes.size();
+		return (int)attributes.size();
 	}
 
 	PointAttribute& operator[](int i) {
